[TS-7361] refactor: usando novo input
